// Basics/FormHandling/handlingts/src/types/types.ts
export interface UserData {
  email: string;
  password: string;
  age: number;
}

// types/types.ts
export interface UserTableType {
  id: number;
  name: string;
  age: number;
}

export interface IUser {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}
export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}
export interface Geo {
  lat: string;
  lng: string;
}
export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}